requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
pydantic-settings
python-dotenv
alembic
beautifulsoup4
readability-lxml
pytest
python-jose[cryptography]
slowapi
